Release candidate 3.2.9
