+++ frontend/src/components/MbfTable.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 // React
 import { useMemo, useRef, useState, memo } from 'react';
 
@@ -73,7 +74,6 @@ function MbfTable<T extends object>({ name, title, columns, rows, getRowKey, foo
 
   const getStableRowKey = (row: T): string | number => {
     if (typeof getRowKey === 'string') {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
       if (row && (row as any)[getRowKey] != null) return (row as any)[getRowKey] as string | number;
     }
     if (typeof getRowKey === 'function') {
@@ -81,7 +81,6 @@ function MbfTable<T extends object>({ name, title, columns, rows, getRowKey, foo
       if (k != null) return k;
     }
     const firstColId = columns?.[0]?.id;
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     if (firstColId && row && (row as any)[firstColId] != null) return (row as any)[firstColId] as string | number;
     console.warn(`MbfTable(${name}): using fallback stable row key; consider providing getRowKey prop for better React performance`);
     let k = rowKeyMapRef.current.get(row);
@@ -92,7 +91,7 @@ function MbfTable<T extends object>({ name, title, columns, rows, getRowKey, foo
     return k;
   };
 
-  // Local states
+  // Local states for column sorting and visibility
   const [orderBy, setOrderBy] = useState<string | null>(localStorage.getItem(`${name}_table_order_by`) || null);
   const [order, setOrder] = useState<'asc' | 'desc' | null>((localStorage.getItem(`${name}_table_order`) as 'asc' | 'desc' | null) || null);
   const [configureVisibilityDialogOpen, setConfigureVisibilityDialogOpen] = useState(false);
@@ -117,6 +116,23 @@ function MbfTable<T extends object>({ name, title, columns, rows, getRowKey, foo
   }, [columns, columnVisibility]);
 
   // Memoized sorted rows
+  // const [sortedRows, setSortedRows] = useState<T[]>(rows);
+  // const workerRef = useRef<Worker>(null);
+  // workerRef.current = new Worker(new URL('./MbfTableSortWorker.tsx', import.meta.url));
+  /*
+  useEffect(() => {
+    if (!workerRef.current) {
+    }
+    const worker = workerRef.current;
+    worker.onmessage = (e) => setSortedRows(e.data);
+    worker.postMessage({ rows, orderBy, order, columns });
+    return () => {
+      worker.onmessage = null;
+    };
+  }, [rows, orderBy, order, columns, name]);  
+  */
+
+  // Memoized sorted rows without web worker
   const sortedRows = useMemo<T[]>(() => {
     if (!orderBy || !order) return rows;
     const sortCol = columns.find((c) => c.id === orderBy);
@@ -129,7 +145,7 @@ function MbfTable<T extends object>({ name, title, columns, rows, getRowKey, foo
       if (typeof sortCol.comparator === 'function') {
         cmp = sortCol.comparator(a.el, b.el);
       } else {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+         
         cmp = comparator<any>(a.el as any, b.el as any, orderBy as string);
       }
       if (cmp !== 0) return order === 'asc' ? cmp : -cmp;
@@ -324,7 +340,6 @@ function MbfTable<T extends object>({ name, title, columns, rows, getRowKey, foo
                   {columns.map((column) => {
                     if (column.hidden) return null;
                     if (!column.required && visibleMap[column.id] === false) return null;
-                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
                     const value = (row as any)[column.id];
                     return (
                       <td
@@ -373,107 +388,6 @@ function MbfTable<T extends object>({ name, title, columns, rows, getRowKey, foo
 }
 
 // Helper to preserve generics with React.memo
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 function typedMemo<T>(c: T): T { return memo(c as any) as T; }
 
-/**
- * MbfTable — data table with sorting, sticky header, and column visibility controls.
- *
- * Summary
- * - Sticky header: Table header cells (`th`) are position: sticky; the parent container
- *   should control scrolling (set a fixed height and `overflow: auto`).
- * - Sorting: Click a header to cycle asc → desc → none. Sort is stable; ties keep input order.
- * - Visibility: Users can show/hide non-required columns via the gear icon dialog.
- * - Persistence: Sort and visibility preferences are stored in `localStorage` using the `name` prop
- *   to form keys: `${name}_table_order_by`, `${name}_table_order`, `${name}_column_visibility`.
- *
- * Row and cell rendering
- * - A cell value is read with `row[column.id]`.
- * - Boolean values render as a disabled checkbox for quick visual status.
- * - If `column.format` is provided and the value is a number, the formatted string is rendered.
- *
- * Sorting details
- * - Comparator rules: null/undefined < numbers (numeric sort) < booleans (false < true) < strings (localeCompare).
- * - Clicking a header toggles: ascending → descending → unsorted (restores original order).
- * - When values are equal, original row order is preserved (stable sort).
- *
- * Layout and scrolling guidance
- * - This component does not enforce its own scrollbars. Place it inside a container that defines
- *   height and `overflow: auto` to enable scrolling while keeping the header sticky.
- *
- * Column definition
- * @typedef {Object} Column
- * @property {string} id
- *   Unique key for the column. Used to read `row[id]`, identify the sort column, and persist
- *   user preferences. Must be stable and unique across the table.
- * @property {string} label
- *   Header text shown in the sticky header for this column.
- * @property {number} [minWidth]
- *   Minimum width in pixels applied to the header and cells. Helps keep columns readable.
- * @property {number} [maxWidth]
- *   Maximum width in pixels. When provided, cells clamp with `white-space: nowrap`,
- *   `overflow: hidden`, and `text-overflow: ellipsis` for graceful truncation.
- * @property {'left'|'center'|'right'} [align='left']
- *   Horizontal alignment for header and body cells.
- * @property {(value: number) => string} [format]
- *   Optional formatter used when the cell value is numeric. Non-numeric values are rendered as-is.
- * @property {boolean} [nosort=false]
- *   If true, disables sorting on this column: header click is ignored, cursor remains default,
- *   and any previously persisted sort on this column is ignored at render time.
- * @property {(value: any, rowKey: string|number, row: Object, column: Column) => import('react').ReactNode} [render]
- *   Optional cell renderer. When provided, it takes precedence over default rendering and
- *   returns JSX for the cell. The `rowKey` is the stable key used for the row, derived from
- *   `getRowKey(row)` or an internal fallback. Note: sorting still uses the raw `row[column.id]`
- *   value, not the rendered output.
- * @property {boolean} [hidden=false]
- *   If true, the column is not rendered initially. Users can still enable it unless `required`.
- * @property {boolean} [required=false]
- *   If true, the column is always visible and cannot be hidden via the dialog.
- *
- * Props
- * @param {Object} props
- * @param {string} props.name
- *   Unique table name used in the UI and to namespace persisted preferences in `localStorage`.
- * @param {MbfTableColumn[]} props.columns
- *   Column configuration array. See Column typedef for all options.
- * @param {Object[]} props.rows
- *   Data rows. Each cell is resolved as `row[column.id]`. For best React performance, provide
- *   a stable identifier per row via `getRowKey`.
- * @param {string | (row: Object) => (string|number)} [props.getRowKey]
- *   Optional string or function that returns a stable key for a row. If omitted, the first column value is
- *   used when available; otherwise an internal stable key is generated for the row object.
- *   Always use useCallback to memoize the getRowKey function:
- *   const getRowKey = useCallback((row) => row.code, []);
- * @param {string} props.footerLeft
- *   Text shown in the left side of the footer bar.
- * @param {string} props.footerRight
- *   Text shown in the right side of the footer bar.
- * @returns {JSX.Element}
- *
- * @example
- * // Using stable keys and selectively updating a single row so only that row re-renders
- * function DevicesTable() {
- *   const [rows, setRows] = useState(initialRows);
- *
- *   return (
- *     <MbTable
- *       name="Devices"
- *       columns={columns}
- *       rows={rows}
- *       getRowKey={(row) => row.code} // stable per-row key
- *     />
- *   );
- * }
- *
- * // Update a specific row (code: F123) to demonstrate selective re-render by key
- * setRows((prev) => {
- *   const idx = prev.findIndex(r => r.code === 'F123');
- *   if (idx === -1) return prev;
- *   const target = prev[idx];
- *   const updated = { ...target, population: (target.population || 0) + 1 };
- *   const next = prev.slice(); // shallow copy of the array container
- *   next[idx] = updated;       // replace only the changed row object
- *   return next;               // other rows keep reference => React skips re-render for them
- * });
- */
 export default typedMemo(MbfTable);
